PROTON-2594: scripts: add PKCS#11 testing preparation script
SoftHSM is an implementation of a cryptographic store accessible through a
PKCS #11 interface. With it, we can test the PKCS#11 functionality
without requiring an actual Hardware Security Module (HSM).

Add a config file for SoftHSM that configures a temporary token dir, an
OpenSSL config that configures use of the PKCS#11 provider, so OpenSSL
is able to process pkcs11: URIs and a script that sets the necessary
environment variables to use these two configs as well as adding a set
of client and server keys and certificates to the SoftHSM, so it can be
consumed later by the pkcs11_test.

Because exported variables are only passed down to children, it's
expected that this script is sourced from the shell that will later on
execute the tests. The working directory should be the source directory
of the project.