PROTON-2594: ci: setup environment for PKCS#11 test
Now that all pieces are in place, let's adapt out Github Action
workflow, so softhsm2, opensc (for pkcs11-tool) and pkcs11-provider
(for OpenSSL integration) are available and amend the ctest step
to source scripts/prep-pkcs11_test.sh, so the relevant environment
variables are exported.

Note that pkcs11-provider is so recent, that we need to update to
ubuntu-24.04 _and_ run apt update, because the used image doesn't
yet know about the package.